Reblocked out the Alarm Clock
Reblocked out the alarm clock ringing. Ask about a camera for each distance or not and if they want more rotation forward and backward not just side to side. Also more violence with the hammer. Also ask about slowing down the squash and stretch so its not the same as the animatic.